// src/render/layout/noto-metrics.gen.ts
export interface FontMetrics {
  unitsPerEm: number;
  ascender: number;
  descender: number;
  advances: Record<400 | 600, Record<number, number>>;
}